Added Weighted HMM implementation

+++ HmmDotNet/HmmDotNet.MachineLearning/Algorithms/Baum-Welch/TimeSensitiveWeightCalculator.cs
@@ -4,19 +4,19 @@ namespace HmmDotNet.MachineLearning.Algorithms.Baum_Welch
 {
     public static class TimeSensitiveWeightCalculator
     {
-        public static double[] Calculate(int k, int T)
+        public static decimal[] Calculate(int k, int T)
         {
-            var result = new double[T];
-            var r = 2d / (k + 1);
+            var result = new decimal[T];
+            var r = 2m / (k + 1);
             for (var t = 0; t < T - 1; t++)
             {
                 if (t <= k - 1)
                 {
-                    result[t] = Math.Pow(1 - r, T - k) * (1d / k);
+                    result[t] = (decimal)Math.Pow(1 -(double)r, T - k) * (1m / k);
                 }
                 else
                 {
-                    result[t] = r * Math.Pow(1 - r, T - t);
+                    result[t] = r * (decimal)Math.Pow(1 - (double)r, T - t);
                 }
             }
             result[T - 1] = r;
